Report LC event status also

+++ numachip2/lc5.h
@@ -28,7 +28,7 @@ public:
 	char name[3];
 
 	void clear(void);
-	uint32_t status(void);
+	uint64_t status(void);
 	LC5(Numachip2& _numachip, const uint16_t _addr, const char *name);
 };
